+++ CHANGELOG.md
@@ -36,6 +36,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- (ci) [fse-478] Updating github actions and versions
+- (ci) [#5](https://github.com/evmos/backend/pull/5) RPC server refactor
+
 ### Improvements
 
 - (ci) [#107](https://github.com/evmos/backend/pull/107) Add golangci linter.